Resolve Code Factor warning
